Move prechecks to checks.go file
